four-model-app/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
transformers==4.44.2
torch>=2.2.0; platform_system!='Windows'
torch==2.2.2+cpu; platform_system=='Windows'
pydantic==2.9.2
python-dotenv==1.0.1
pytest==8.3.2
httpx==0.27.2
python-whois==0.9.4
tldextract==5.1.2
requests==2.32.3
